ANNDEA v3: MCTr1

+++ Code/MCTr1_GenerateTrainClassifyTracks.py
@@ -153,7 +153,7 @@ if os.path.isfile(required_file_location)==False:
             data_aggregated=data.groupby(['Rec_Seg_ID'])['z'].min().reset_index()
             data_aggregated_show=data_aggregated.groupby(['z']).count().reset_index()
             data_aggregated_show=data_aggregated_show.rename(columns={'Rec_Seg_ID': "No_Tracks"})
-            data_aggregated_show['PID']=data_aggregated_show['z'].rank(ascending=True)
+            data_aggregated_show['PID']=data_aggregated_show['z'].rank(ascending=True).astype(int)
             print(data_aggregated_show)
             exit()
